requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1